Move nb/serverextension tests to own file
