+++ integration-tests/test_hub.py
@@ -2,7 +2,6 @@ import requests
 from hubtraf.user import User
 from hubtraf.auth.dummy import login_dummy
 import secrets
-import subprocess
 import pytest
 from functools import partial
 import asyncio
@@ -106,44 +105,3 @@ async def test_user_admin_remove():
 
             # Assert that the user does *not* have admin rights
             assert f'jupyter-{username}' in grp.getgrnam('jupyterhub-admins').gr_mem
-
-
-def test_serverextensions():
-    """
-    Validate serverextensions we want are installed
-    """
-    # jupyter-serverextension writes to stdout and stderr weirdly
-    proc = subprocess.run([
-        '/opt/tljh/user/bin/jupyter-serverextension',
-        'list', '--sys-prefix'
-    ], stderr=subprocess.PIPE)
-
-    extensions = [
-        'jupyterlab 0.32.1',
-        'nbgitpuller 0.6.1',
-        'nteract_on_jupyter 1.8.1',
-        'nbresuse '
-    ]
-
-    for e in extensions:
-        assert '{} \x1b[32mOK\x1b[0m'.format(e) in proc.stderr.decode()
-
-def test_nbextensions():
-    """
-    Validate nbextensions we want are installed & enabled
-    """
-    # jupyter-nbextension writes to stdout and stderr weirdly
-    proc = subprocess.run([
-        '/opt/tljh/user/bin/jupyter-nbextension',
-        'list', '--sys-prefix'
-    ], stderr=subprocess.PIPE, stdout=subprocess.PIPE)
-
-    extensions = [
-        'nbresuse/main',
-    ]
-
-    for e in extensions:
-        assert '{} \x1b[32m enabled \x1b[0m'.format(e) in proc.stdout.decode()
-
-    # Ensure we have 'OK' messages in our stdout, to make sure everything is importable
-    proc.stderr.decode() == '      - Validating: \x1b[32mOK\x1b[0m\n' * len(extensions)
